Solidify UX, with modal for new expense form. Prepared camera view and photo lib views for picture selection

+++ ExpensesBot/Models/ExpenseModel.swift
@@ -13,14 +13,14 @@ final class ExpenseModel {
     @Attribute(.unique) var id: UUID
     @Attribute(.externalStorage) var imageData: Data
     var timestamp: Date
-    var total: Float
+    var total: Double
     var currency: String
     var expenseDescription: String
 
     init(id: UUID,
          imageData: Data,
          timestamp: Date,
-         total: Float,
+         total: Double,
          currency: String,
          expenseDescription: String) {
         self.id = id
